Avoid warning for changing copy of dataframe
Pandas warns that this assignment may or may not change the
original data frame, because it can't guarantee whether
matching_runs is a copy or a view. As far as I can tell, it should
always be a copy, but just in case, and to avoid the warning,
create a new dataframe here.
Note that this only occurs when the state we are aligning to is
observed at the first epoch, otherwise the drop() will fail anyway.