Work with raw data instead of aggregated runs
Working with the runs fails when we have a sequence such as
Awake -> REM -> Awake. In those cases, the middle section will be
identified as both Awake and REM, resulting in two overlapping
boxes.
Plotting the raw data makes life easier: no need to duplicate the
last time/state, and also avoids the problem above. This should
still be fine if we work with aligned data in the future.