Reset internals on new alignment
This isn't something that would happen in current usage, but it came up
in another test. If an experiment was aligned twice, the new alignment
runs were appended to the first ones instead of overwriting them. This
fixes the bug and alters a relevant test to check for any regression.