Don't line up stimulus occurrences with state runs
The tests for alignment to stimulus were not strong enough to
catch a bug in get_alignment_data. This way should make it clearer
to see if the right subsequence of states is retrieved.
Also deleting the Excel file for these tests, as it was never used
and editing the CSV is easier.